Revert "Added like by location and fixed account validation checks"
