--- src/instapy2-bak/__init__.py
@@ -1,1 +1,0 @@
-from .instapy2 import InstaPy2
