--- src/instapy2-bak/types/__init__.py
@@ -1,5 +1,0 @@
-from .comment_type import CommentType
-from .like_type import LikeType
-from .follow_type import FollowType
-from .post_type import PostType
-from .unfollow_type import UnfollowType
