--- src/instapy2-bak/utilities/__init__.py
@@ -1,6 +1,0 @@
-from .comments import CommentsUtility
-from .follows import FollowsUtility
-from .interactions import InteractionsUtility
-from .likes import LikesUtility
-from .limitations import LimitationsUtility
-from .messages import MessageUtility
