+++ src/instapy2/helpers/__init__.py
@@ -1,0 +1,2 @@
+from .location import LocationHelper
+from .people import PeopleHelper
