+++ src/instapy2/types/__init__.py
@@ -1,2 +1,5 @@
-from .commenttype import CommentType
-from .liketype import LikeType
+from .comment_type import CommentType
+from .like_type import LikeType
+from .follow_type import FollowType
+from .post_type import PostType
+from .unfollow_type import UnfollowType
