--- src/instapy2/types/commenttype.py
@@ -1,5 +1,0 @@
-from enum import Enum
-
-class CommentType(Enum):
-    HASHTAG = 0
-    USER = 1
