--- src/instapy2/types/liketype.py
@@ -1,6 +1,0 @@
-from enum import Enum
-
-class LikeType(Enum):
-    HASHTAG = 0
-    LOCATION = 1
-    USER = 2
