+++ src/instapy2/utilities/__init__.py
@@ -1,2 +1,6 @@
-from .logger import Logger
-from .utility import Utility
+from .comments import CommentsUtility
+from .follows import FollowsUtility
+from .interactions import InteractionsUtility
+from .likes import LikesUtility
+from .limitations import LimitationsUtility
+from .messages import MessageUtility
