+++ src/instapy2/utilities/limitations.py
@@ -1,83 +1,37 @@
 from instagrapi import Client
-from instagrapi.types import Media
+from instagrapi.types import Media, UserShort
 
-class Limitations:
+from typing import Dict, Tuple
+
+class LimitationsUtility:
     def __init__(self, session: Client):
         self.session = session
 
-        self.commenter_range = (1, 1000)
+        self.commenters_range = (1, 100)
         self.enabled = False
-        self.follower_range = (1, 1000)
-        self.following_range = (1, 1000)
+        self.followers_range = (1, 1000)
 
-        self.skip_business = False
-        self.skip_private = True
-        self.skip_verified = False
+    def from_json(self, data: Dict):
+        self.commenters_range = (data['commenters_range'][0] or 1, data['commenters_range'][1] or 100)
+        self.enabled = data['enabled'] or False
+        self.followers_range = (data['followers_range'][0] or 1, data['followers_range'][1] or 1)
 
-
-    def set_commenter_range(self, range: tuple[int, int]):
-        self.commenter_range = range
+    def set_commenters_range(self, range: Tuple[int, int]):
+        self.commenters_range = range
 
     def set_enabled(self, enabled: bool):
         self.enabled = enabled
 
-    def set_follower_range(self, range: tuple[int, int]):
-        self.follower_range = range
-
-    def set_skip_business(self, skip: bool):
-        self.skip_business = skip
-
-    def set_skip_private(self, skip: bool):
-        self.skip_private = skip
-
-    def set_skip_verified(self, skip: bool):
-        self.skip_verified = skip
+    def set_followers_range(self, range: Tuple[int, int]):
+        self.followers_range = range
 
 
-    def is_account_appropriate(self, media: Media) -> bool:
-        enabled = self.enabled
-        username = media.user.username
-
-        if not enabled:
-            return True
-        else:
-            if username is None:
-                return False
-            else:
-                try:
-                    user_info = self.session.user_info_by_username(username=username)
-                except:
-                    user_info = None
-                
-                return False if user_info is None else all([self.skip_business and not user_info.is_business, self.skip_private and not user_info.is_private, self.skip_verified and not user_info.is_verified])
-
-    def within_commenter_range(self, media: Media) -> bool:
-        enabled = self.enabled
-        min, max = self.commenter_range
-        
-        commenter_count = media.comment_count or 0
-        return True if not enabled else min <= commenter_count <= max
-
-    def within_follower_range(self, media: Media) -> bool:
-        enabled = self.enabled
-        min, max = self.follower_range
-        username = media.user.username
-
-        try:
-            follower_count = self.session.user_info_by_username(username=username).follower_count if username is not None else 0
-        except:
-            follower_count = 0
-        
-        return True if not enabled else min <= follower_count <= max
+    def within_commenters_range(self, media: Media) -> bool:
+        min, max = self.commenters_range
+        commenters_count = media.comment_count
+        return self.enabled if self.enabled else min <= commenters_count if commenters_count is not None else 0 <= max
     
-    def within_following_range(self, media: Media) -> bool:
-        enabled = self.enabled
-        min, max = self.following_range
-        username = media.user.username
-
-        try:
-            following_count = self.session.user_info_by_username(username=username).following_count if username is not None else 0
-        except:
-            following_count = 0
-        
-        return True if not enabled else min <= following_count <= max
+    def within_followers_range(self, user: UserShort) -> bool:
+        min, max = self.followers_range
+        followers_count = self.session.user_info(user_id=self.session.user_id_from_username(username=user.username if user.username is not None else '')).follower_count
+        return self.enabled if self.enabled else min <= followers_count <= max
