+++ src/instapy2/utilities/messages.py
@@ -17,12 +17,29 @@ class MessageUtility:
         self.percentage = data['percentage'] or 0
 
     def set_enabled(self, enabled: bool):
+        """
+            Enables the ability to direct message a user.
+            
+            :param enabled: enabled=True means a user will be direct messaged.
+        """
         self.enabled = enabled
 
     def set_messages(self, messages: List[str]):
+        """
+            Sets the messages to be used when direct messaging a user.
+
+            :param messages: messages=['message 1', ' message 2', 'message 3 @{}']
+            
+            Adding @{} will tag the other user in the direct message.
+        """
         self.messages = messages
 
     def set_percentage(self, percentage: int):
+        """
+            Set the percentage of users to direct message.
+
+            :param percentage: percentage=25 means every 4th user will be direct messaged.
+        """
         self.percentage = percentage
 
     def message(self, user: Union[str, UserShort], text: str) -> Tuple[Union[Exception, None], bool]:
